Moved quote table columns def into quotes page
