+++ src/pages/Quotes/Quotes.jsx
@@ -3,44 +3,38 @@ import {
   Box,
   Flex,
   useToast,
-  HStack,
   Tooltip,
-  FormControl,
   Input,
   Text,
   useDisclosure,
   VStack,
-  IconButton,
-  Icon,
-  Card,
-  CardBody,
-  Skeleton,
   InputGroup,
   InputLeftElement,
   Select,
-  useColorModeValue
+  useColorModeValue,
+  Badge,
+  Button,
+  Avatar
 } from '@chakra-ui/react';
-import { MdPostAdd, MdSearch, MdFilterList, MdFilterAlt } from 'react-icons/md';
 import {
-  QuoteTable,
   ConnectedQuoteDeleteAlertDialog,
   CreateQuoteForm,
   EditQuoteForm,
   PageHeader,
   QuoteStatCards
 } from '../../components';
-import { TbRuler } from 'react-icons/tb';
-import { useFetchQuotes, useSearchQuote, useUpdateQuote } from '../../hooks/useAPI/useQuotes';
+import { useFetchQuotes, useUpdateQuote } from '../../hooks/useAPI/useQuotes';
 import { useQuoteStatuses } from '../../hooks/useAPI/useQuoteStatuses';
-// import { useQueryClient } from '@tanstack/react-query';
-import useDebounce from '../../hooks/useDebounce';
 import { useFetchAllServices } from '../../hooks/useAPI/useServices';
-import { Circle, Search } from 'lucide-react';
+import { ChevronRight, Pencil, Trash, Search } from 'lucide-react';
 import DataTable from '../../components/ui/DataTable';
-import quoteColumns from '../../components/Quotes/Tables/QuoteColumns';
+import { createColumnHelper } from '@tanstack/react-table';
+import { formatMoneyValue, monthDDYYYYFormat, formatNumber } from '../../utils';
+import { Link } from 'react-router-dom';
 
 function Estimates() {
   // const queryClient = useQueryClient();
+  const columnHelper = createColumnHelper();
 
   // Chakra UI Hooks
   const initialRef = React.useRef();
@@ -51,16 +45,10 @@ function Estimates() {
 
   // Custom React Hooks
   const { quotes, isLoading: quotesLoadingStateIsOn, isError } = useFetchQuotes();
-  const {
-    data: services,
-    isLoading: isRoofingServicesLoading,
-    isError: isRoofingServicesError
-  } = useFetchAllServices();
+  const { data: services, isLoading: isRoofingServicesLoading } = useFetchAllServices();
   const { quoteStatuses } = useQuoteStatuses();
 
   // States to manage data
-  const [selectedEstimateId, setSelectedEstimateId] = useState('');
-  const [selectedEstimateNumber, setSelectedEstimateNumber] = useState('');
   const [selectedEditQuote, setSelectedEditQuote] = useState({
     id: '',
     quote_number: '',
@@ -73,26 +61,185 @@ function Estimates() {
     measurement_note: '',
     cust_note: ''
   });
+  const [selectedRowItem, setSelectedRowItem] = useState('');
 
   const { mutate: mutateUpdateQuote, isLoading: quoteUpdateIsLoading } = useUpdateQuote(toast);
+  // const deboundedQuoteSearchTerm = useDebounce(searchQuoteInput, 100);
 
-  const [searchQuote, setSearchQuote] = useState('');
-  const [searchQuoteInput, setSearchQuoteInput] = useState('');
-  const deboundedQuoteSearchTerm = useDebounce(searchQuoteInput, 100);
-  const { quoteSearchResult, quoteSearchIsLoading } = useSearchQuote(deboundedQuoteSearchTerm);
+  const quoteColumns = [
+    columnHelper.accessor('quote_number', {
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Text
+            fontWeight={500}
+            fontSize={'14px'}
+            textColor={useColorModeValue('gray.900', 'gray.200')}>
+            #{formatNumber(quote.quote_number)}
+          </Text>
+        );
+      },
+      header: () => <Text>Quote</Text>
+    }),
+    columnHelper.accessor('quote_date', {
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Text fontWeight={400} fontSize={'14px'}>
+            {monthDDYYYYFormat(quote.quote_date)}
+          </Text>
+        );
+      },
+      header: () => <Text>Date</Text>
+    }),
+    columnHelper.accessor('status_id', {
+      header: () => <Text>Status</Text>,
+      cell: ({ row }) => {
+        const quote = row.original;
+        if (quote.quote_status.name === 'Accepted') {
+          return (
+            <Badge rounded={'full'} py={1} px={2} colorScheme="green">
+              <Flex gap={2}>
+                <Box w={1} h={1} p={1} bg={'green.500'} rounded={'full'} my={'auto'}></Box>
+                <Text textTransform={'initial'} fontSize={'12px'} fontWeight={500}>
+                  {quote.quote_status.name}
+                </Text>
+              </Flex>
+            </Badge>
+          );
+        } else if (quote.quote_status.name === 'Pending') {
+          return (
+            <Badge rounded={'full'} py={1} px={2} colorScheme="yellow">
+              <Flex gap={2}>
+                <Box w={1} h={1} p={1} bg={'yellow.500'} rounded={'full'} my={'auto'}></Box>
+                <Text textTransform={'initial'} fontSize={'12px'} fontWeight={500}>
+                  {quote.quote_status.name}
+                </Text>
+              </Flex>
+            </Badge>
+          );
+        } else if (quote.quote_status.name === 'Rejected') {
+          return (
+            <Badge rounded={'full'} py={1} px={2} colorScheme="red">
+              <Flex gap={2}>
+                <Box w={1} h={1} p={1} bg={'red.500'} rounded={'full'} my={'auto'}></Box>
+                <Text textTransform={'initial'} fontSize={'12px'} fontWeight={500}>
+                  {quote.quote_status.name}
+                </Text>
+              </Flex>
+            </Badge>
+          );
+        } else {
+          return (
+            <Badge rounded={'full'} py={1} px={2} colorScheme="gray">
+              <Flex gap={2}>
+                <Box w={1} h={1} p={1} bg={'gray.500'} rounded={'full'} my={'auto'}></Box>
+                <Text textTransform={'initial'} fontSize={'12px'} fontWeight={500}>
+                  {quote.quote_status.name}
+                </Text>
+              </Flex>
+            </Badge>
+          );
+        }
+      }
+    }),
+    columnHelper.accessor('service', {
+      id: 'service',
+      header: () => <Text>Service</Text>,
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Text fontWeight={400} fontSize={'14px'}>
+            {quote.services.name}
+          </Text>
+        );
+      }
+    }),
+    columnHelper.accessor('expiration_date', {
+      header: () => <Text>Expiration</Text>,
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Text fontWeight={400} fontSize={'14px'}>
+            {monthDDYYYYFormat(quote.expiration_date)}
+          </Text>
+        );
+      }
+    }),
+    columnHelper.accessor((row) => `${row.customer.first_name}`, {
+      id: 'customer',
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Link
+            to={`/customers/${quote.customer.id}`}
+            _hover={{ bg: useColorModeValue('gray.200', 'gray.600') }}>
+            <Button variant={'ghost'}>
+              <Flex gap={3}>
+                <Avatar
+                  my={'auto'}
+                  size={'sm'}
+                  name={`${quote.customer.first_name} ${quote.customer.last_name}`}
+                  bg={useColorModeValue('gray.200', 'gray.600')}
+                  textColor={useColorModeValue('gray.700', 'gray.200')}
+                />
+                <Box fontSize={'14px'}>
+                  <Flex gap={1} fontWeight={500}>
+                    <Text>{quote.customer.first_name}</Text>
+                    <Text>{quote.customer.last_name}</Text>
+                  </Flex>
+                  <Text fontWeight={400}>{quote.customer.email}</Text>
+                </Box>
+              </Flex>
+            </Button>
+          </Link>
+        );
+      }
+    }),
+    columnHelper.accessor('total', {
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Text fontSize={'14px'} fontWeight={400}>
+            ${formatMoneyValue(quote.total)}
+          </Text>
+        );
+      }
+    }),
+    columnHelper.accessor('actions', {
+      header: () => <Text>Actions</Text>,
+      cell: ({ row }) => {
+        const quote = row.original;
+        return (
+          <Flex gap={2}>
+            <Tooltip label="Edit">
+              <Button p={0} onClick={() => handleEditDrawer(quote)}>
+                <Pencil size={'15px'} />
+              </Button>
+            </Tooltip>
+            <Tooltip label="Delete">
+              <Button p={0} onClick={() => deleteModalHandler(quote.quote_number)}>
+                <Trash size={'15px'} />
+              </Button>
+            </Tooltip>
+            <Tooltip label="Quote Details">
+              <Link to={`/quotes/${quote.quote_number}`}>
+                <Button px={0}>
+                  <ChevronRight size={'15px'} />
+                </Button>
+              </Link>
+            </Tooltip>
+          </Flex>
+        );
+      }
+    })
+  ];
 
-  const searchEstimate = async (e) => {
-    e.preventDefault();
-  };
-
-  // Update this function to handle delete functinality.
-  const handleDelete = (estimateId, estimate_number) => {
-    setSelectedEstimateId(estimateId);
-    setSelectedEstimateNumber(estimate_number);
+  const deleteModalHandler = (quote_number) => {
+    setSelectedRowItem(quote_number);
     onDeleteOpen();
   };
 
-  /////////////////////////// Functions that handle edit functionality ////////////////////////////////
   const handleEditDrawer = (quote) => {
     setSelectedEditQuote({
       id: quote.id,
@@ -147,9 +294,9 @@ function Estimates() {
         onClose={onDeleteClose}
         onOpen={onDeleteOpen}
         toast={toast}
-        itemNumber={selectedEstimateNumber}
+        itemNumber={selectedRowItem}
         header={`Delete Quote`}
-        entityDescription={`Quote #${selectedEstimateNumber}`}
+        entityDescription={`Quote #${selectedRowItem}`}
         body={`Once you confirm you can't undo this action afterwards. There will be no way to restore the information. 🚨`}
       />
       <EditQuoteForm
@@ -163,7 +310,7 @@ function Estimates() {
         handleEditSubmit={handleEditSubmit}
         loadingState={quoteUpdateIsLoading}
       />
-      <VStack my={'4'} w="full" mx={'auto'} px={{ base: '4', lg: '8' }} gap={4}>
+      <VStack mt={'4'} mb={10} w="full" mx={'auto'} px={{ base: '4', lg: '8' }} gap={4}>
         <PageHeader
           title={'Quotes'}
           subheading={'Manage your quotes to send out to your customers.'}
